LU-9019 ptlrpc: migrate pinger to 64 bit time
Replace cfs_time_current_sec() to avoid the overflow
issues in 2038 with ktime_get_real_seconds(). Remove
cfs_timeout_cap() and CFS_DURATION abstraction. Change
imp_next_ping, obd_eviction_timer, and ti_timeout to
time64_t. With these changes the pinger will be 64
bit time compliant.

Reviewed-on: https://review.whamcloud.com/28035
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ lustre/ptlrpc/recover.c
@@ -352,9 +352,9 @@ int ptlrpc_recover_import(struct obd_import *imp, char *new_uuid, int async)
 
         if (!async) {
                 struct l_wait_info lwi;
-                int secs = cfs_time_seconds(obd_timeout);
+		long secs = cfs_time_seconds(obd_timeout);
 
-                CDEBUG(D_HA, "%s: recovery started, waiting %u seconds\n",
+		CDEBUG(D_HA, "%s: recovery started, waiting %lu seconds\n",
                        obd2cli_tgt(imp->imp_obd), secs);
 
                 lwi = LWI_TIMEOUT(secs, NULL, NULL);
